+++ Source/XYTOR/Private/Core/Dialogues/PS_Dialogues.cpp
@@ -1,0 +1,14 @@
+// XYTOR. All Rights Reserved.
+
+
+#include "Core/Dialogues/PS_Dialogues.h"
+
+APS_Dialogues::APS_Dialogues()
+{
+    DialogueComponent = CreateDefaultSubobject<UAC_Dialogues>(TEXT("Dialogue"));
+}
+
+UAC_Dialogues* APS_Dialogues::GetDialogue() const
+{
+    return DialogueComponent;
+}
