+++ Source/XYTOR/Public/Core/Weapons/PC_Weapons.h
@@ -3,16 +3,11 @@
 #pragma once
 
 #include "CoreMinimal.h"
-#include "GameFramework/PlayerController.h"
+#include "Core/Detective/PC_Detective.h"
 #include "PC_Weapons.generated.h"
 
-/**
- * 
- */
 UCLASS()
-class XYTOR_API APC_Weapons : public APlayerController
+class XYTOR_API APC_Weapons : public APC_Detective
 {
     GENERATED_BODY()
-
-protected:
 };
